Extract LedStrip as an interface
and implements it with Apa102LedStrip
